+++ jest.framework.config.cjs
@@ -1,5 +1,5 @@
 module.exports = {
   testEnvironment: 'node',
-  testMatch: ['**/__tests__/framework/**/*.test.js'],
+  testMatch: ['**/__tests__/unit/**/*.test.js'],
   testPathIgnorePatterns: ['/node_modules/', '/bin/'],
 };